Fix coalesce bug with large Cartesian product sample spaces.
Basically, we need to obtain the outcome class during
construction, but itertools.product() will consume the iterables.
If the alphabets for the new objects are themselves Cartesian
products, then it would fail to consume. Now, we determine
the outcome class more intelligently.

This also addresses an issue when comparing CartesianProducts,
abc.Set would defer to __le__ which consumes the CartesianProduct.
Now, we just compare the alphabets directly.